B2B-1876: New directories in pub/media do not get synced back to local filesystem from S3 Remote Storage

+++ app/code/Magento/Cms/Model/Wysiwyg/Images/Storage.php
@@ -9,6 +9,7 @@ declare(strict_types=1);
 namespace Magento\Cms\Model\Wysiwyg\Images;
 
 use Magento\Cms\Helper\Wysiwyg\Images;
+use Magento\Framework\App\Config\ScopeConfigInterface;
 use Magento\Framework\App\Filesystem\DirectoryList;
 use Magento\Framework\App\ObjectManager;
 
@@ -34,6 +35,9 @@ class Storage extends \Magento\Framework\DataObject
     const THUMBS_DIRECTORY_NAME = '.thumbs';
 
     const THUMB_PLACEHOLDER_PATH_SUFFIX = 'Magento_Cms::images/placeholder_thumbnail.jpg';
+
+    private const MEDIA_GALLERY_IMAGE_FOLDERS_CONFIG_PATH
+        = 'system/media_storage_configuration/allowed_resources/media_gallery_image_folders';
 
     /**
      * Config object
@@ -159,6 +163,26 @@ class Storage extends \Magento\Framework\DataObject
     private $mime;
 
     /**
+     * @var ScopeConfigInterface
+     */
+    private $coreConfig;
+
+    /**
+     * @var string
+     */
+    private $allowedPathPattern;
+
+    /**
+     * @var string
+     */
+    private $allowedTargetDirPattern;
+
+    /**
+     * @var array
+     */
+    private $allowedDirs;
+
+    /**
      * Construct
      *
      * @param \Magento\Backend\Model\Session $session
@@ -181,6 +205,7 @@ class Storage extends \Magento\Framework\DataObject
      * @param \Magento\Framework\Filesystem\Io\File|null $ioFile
      * @param \Psr\Log\LoggerInterface|null $logger
      * @param \Magento\Framework\File\Mime $mime
+     * @param ScopeConfigInterface $coreConfig
      *
      * @throws \Magento\Framework\Exception\FileSystemException
      * @SuppressWarnings(PHPMD.ExcessiveParameterList)
@@ -205,7 +230,8 @@ class Storage extends \Magento\Framework\DataObject
         \Magento\Framework\Filesystem\DriverInterface $file = null,
         \Magento\Framework\Filesystem\Io\File $ioFile = null,
         \Psr\Log\LoggerInterface $logger = null,
-        \Magento\Framework\File\Mime $mime = null
+        \Magento\Framework\File\Mime $mime = null,
+        ScopeConfigInterface $coreConfig = null
     ) {
         $this->_session = $session;
         $this->_backendUrl = $backendUrl;
@@ -226,6 +252,7 @@ class Storage extends \Magento\Framework\DataObject
         $this->file = $file ?: ObjectManager::getInstance()->get(\Magento\Framework\Filesystem\Driver\File::class);
         $this->ioFile = $ioFile ?: ObjectManager::getInstance()->get(\Magento\Framework\Filesystem\Io\File::class);
         $this->mime = $mime ?: ObjectManager::getInstance()->get(\Magento\Framework\File\Mime::class);
+        $this->coreConfig = $coreConfig ?: ObjectManager::getInstance()->get(ScopeConfigInterface::class);
         parent::__construct($data);
     }
 
@@ -254,6 +281,7 @@ class Storage extends \Magento\Framework\DataObject
      * Prepare and get conditions for exclude directories
      *
      * @return array
+     * @deprecated
      */
     protected function getConditionsForExcludeDirs()
     {
@@ -281,6 +309,7 @@ class Storage extends \Magento\Framework\DataObject
      * @param \Magento\Framework\Data\Collection\Filesystem $collection
      * @param array $conditions
      * @return \Magento\Framework\Data\Collection\Filesystem
+     * @deprecated
      */
     protected function removeItemFromCollection($collection, $conditions)
     {
@@ -318,9 +347,11 @@ class Storage extends \Magento\Framework\DataObject
             ->setCollectRecursively(false)
             ->setOrder('basename', \Magento\Framework\Data\Collection\Filesystem::SORT_ORDER_ASC);
 
-        $conditions = $this->getConditionsForExcludeDirs();
+        if (preg_match($this->getAllowedTargetDirPattern(), $path) != 1) {
+            $collection->setDirsFilter($this->getAllowedDirMask($path));
+        }
 
-        return $this->removeItemFromCollection($collection, $conditions);
+        return $collection;
     }
 
     /**
@@ -430,6 +461,12 @@ class Storage extends \Magento\Framework\DataObject
      */
     public function createDirectory($name, $path)
     {
+        if (!($this->isPathAllowed($path) || preg_match($this->getAllowedTargetDirPattern(), $path))) {
+            throw new \Magento\Framework\Exception\LocalizedException(
+                __('We cannot create folder %1 under %2', $name, $this->_getRelativePathToRoot($path))
+            );
+        }
+
         if (!preg_match(self::DIRECTORY_NAME_REGEXP, $name)) {
             throw new \Magento\Framework\Exception\LocalizedException(
                 __('Please rename the folder using only Latin letters, numbers, underscores and dashes.')
@@ -477,14 +514,16 @@ class Storage extends \Magento\Framework\DataObject
      */
     public function deleteDirectory($path)
     {
-        if ($this->_coreFileStorageDb->checkDbUsage()) {
-            $this->_directoryDatabaseFactory->create()->deleteDirectory($path);
-        }
-        if (!$this->isPathAllowed($path, $this->getConditionsForExcludeDirs())) {
+        if (!$this->isPathAllowed($path)) {
             throw new \Magento\Framework\Exception\LocalizedException(
                 __('We cannot delete directory %1.', $this->_getRelativePathToRoot($path))
             );
         }
+
+        if ($this->_coreFileStorageDb->checkDbUsage()) {
+            $this->_directoryDatabaseFactory->create()->deleteDirectory($path);
+        }
+
         try {
             $this->_deleteByPath($path);
             $path = $this->getThumbnailRoot() . $this->_getRelativePathToRoot($path);
@@ -524,6 +563,11 @@ class Storage extends \Magento\Framework\DataObject
      */
     public function deleteFile($target)
     {
+        if (!$this->isPathAllowed($target)) {
+            throw new \Magento\Framework\Exception\LocalizedException(
+                __('We can\'t delete the file right now.')
+            );
+        }
         $relativePath = $this->_directory->getRelativePath($target);
         if ($this->_directory->isFile($relativePath)) {
             $this->_directory->delete($relativePath);
@@ -551,7 +595,7 @@ class Storage extends \Magento\Framework\DataObject
      */
     public function uploadFile($targetPath, $type = null)
     {
-        if (!$this->isPathAllowed($targetPath, $this->getConditionsForExcludeDirs())) {
+        if (!($this->isPathAllowed($targetPath) || preg_match($this->getAllowedTargetDirPattern(), $targetPath))) {
             throw new \Magento\Framework\Exception\LocalizedException(
                 __('We can\'t upload the file to current folder right now. Please try another folder.')
             );
@@ -903,27 +947,118 @@ class Storage extends \Magento\Framework\DataObject
     }
 
     /**
-     * Check if path is not in excluded dirs.
+     * Check if path is allowed
      *
      * @param string $path Absolute path
-     * @param array $conditions Exclude conditions
      * @return bool
      */
-    private function isPathAllowed($path, array $conditions): bool
+    private function isPathAllowed($path): bool
     {
-        $isAllowed = true;
-        $regExp = $conditions['reg_exp'] ? '~' . implode('|', array_keys($conditions['reg_exp'])) . '~i' : null;
+        return preg_match($this->getAllowedPathPattern(), $path) == 1;
+    }
+
+    /**
+     * Get allowed path pattern
+     *
+     * @return string
+     */
+    private function getAllowedPathPattern()
+    {
+        if (null === $this->allowedPathPattern) {
+            $mediaGalleryImageFolders = $this->coreConfig->getValue(
+                self::MEDIA_GALLERY_IMAGE_FOLDERS_CONFIG_PATH,
+                'default'
+            );
+            $regExp = '~';
+            $or = '';
+            foreach($mediaGalleryImageFolders as $folder) {
+                $folderPattern = str_replace('/', '[/\\\]+', $folder);
+                $regExp .= $or . 'media[/\\\]+' . $folderPattern . '[/\\\]+[a-z0-9\.\-\_]+';
+                $or = '|';
+            }
+            $regExp .= '~i';
+            $this->allowedPathPattern = $regExp;
+        }
+        return $this->allowedPathPattern;
+    }
+
+    /**
+     * Get allowed target dir pattern
+     *
+     * @return string
+     */
+    private function getAllowedTargetDirPattern()
+    {
+        if (null === $this->allowedTargetDirPattern) {
+            $mediaGalleryImageFolders = $this->coreConfig->getValue(
+                self::MEDIA_GALLERY_IMAGE_FOLDERS_CONFIG_PATH,
+                'default'
+            );
+            $regExp = '~';
+            $or = '';
+            foreach($mediaGalleryImageFolders as $folder) {
+                $folderPattern = str_replace('/', '[/\\\]+', $folder);
+                $regExp .= $or . 'media[/\\\]+' . $folderPattern . '[/\\\]*$';
+                $or = '|';
+            }
+            $regExp .= '~i';
+            $this->allowedTargetDirPattern = $regExp;
+        }
+        return $this->allowedTargetDirPattern;
+    }
+
+    /**
+     * Get allowed media gallery image folders
+     *
+     * example:
+     *   [
+     *     [0 => 'wysiwyg'],
+     *     [0 => 'catalog', 1 => 'category']
+     *   ];
+     *
+     * @return array
+     */
+    private function getAllowedDirs(): array
+    {
+        if (null == $this->allowedDirs) {
+            $imageFolders = $this->coreConfig->getValue(
+                self::MEDIA_GALLERY_IMAGE_FOLDERS_CONFIG_PATH,
+                'default'
+            );
+
+            $this->allowedDirs = [];
+            foreach ($imageFolders as $folder) {
+                $this->allowedDirs[] = explode('/', $folder);
+            }
+        }
+        return $this->allowedDirs;
+    }
+
+    /**
+     * Get allowed dir mask.
+     *
+     * @param string $path
+     * @return string
+     */
+    private function getAllowedDirMask(string $path)
+    {
+        $allowedDirs = $this->getAllowedDirs();
+        $column = 0;
         $storageRoot = $this->_cmsWysiwygImages->getStorageRoot();
         $storageRootLength = strlen($storageRoot);
-
         $mediaSubPathname = substr($path, $storageRootLength);
-        $rootChildParts = explode('/', '/' . ltrim($mediaSubPathname, '/'));
-
-        if (array_key_exists($rootChildParts[1], $conditions['plain'])
-            || ($regExp && preg_match($regExp, $path))) {
-            $isAllowed = false;
+        if ($mediaSubPathname) {
+            $pathSegments = explode('/', ltrim($mediaSubPathname, '/'));
+            foreach ($pathSegments as $index => $pathSegment) {
+                $subDirKeys = array_keys(array_column($allowedDirs, $index), $pathSegment);
+                $dirs = [];
+                foreach ($subDirKeys as $subDirKey) {
+                    $dirs[] = $allowedDirs[$subDirKey];
+                }
+                $allowedDirs = $dirs;
+                $column++;
+            }
         }
-
-        return $isAllowed;
+        return '/^(' . implode('|', array_unique(array_column($allowedDirs, $column))) . ')$/';
     }
 }
